src/test/java/project/minju1/service/MemberServiceTest.java
package project.minju1.service;

import org.assertj.core.api.Assertions;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.transaction.annotation.Transactional;
import project.minju1.entity.Member;
import project.minju1.repository.MemberRepository;

import javax.persistence.EmbeddedId;
import javax.persistence.EntityManager;
import javax.persistence.PersistenceContext;

import java.util.List;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.api.Assertions.*;

@SpringBootTest
@Transactional
class MemberServiceTest {

    @Autowired
    MemberRepository memberRepository;

    @Autowired
    MemberService memberService;

    @BeforeEach
    public void before(){

    }

    @Test
    public void serviceTest(){

        Member member1 = new Member("m1");
        Long member1Id = memberService.join(member1);

        assertThat(member1Id).isEqualTo(member1.getId());
    }

}